+++ src/main/java/edu/illinois/library/cantaloupe/status/ApplicationStatus.java
@@ -3,8 +3,6 @@ package edu.illinois.library.cantaloupe.status;
 import edu.illinois.library.cantaloupe.Application;
 import edu.illinois.library.cantaloupe.cache.InfoCache;
 import edu.illinois.library.cantaloupe.cache.InfoService;
-import edu.illinois.library.cantaloupe.script.DelegateProxyService;
-import edu.illinois.library.cantaloupe.script.InvocationCache;
 
 import java.lang.management.ManagementFactory;
 import java.lang.management.RuntimeMXBean;
@@ -19,22 +17,6 @@ public final class ApplicationStatus {
 
     public String getApplicationVersion() {
         return Application.getVersion();
-    }
-
-    /**
-     * @return Max delegate method invocation cache size.
-     */
-    public long getDMICMaxSize() {
-        InvocationCache cache = DelegateProxyService.getInvocationCache();
-        return cache.maxSize();
-    }
-
-    /**
-     * @return Current delegate method invocation cache size.
-     */
-    public long getDMICSize() {
-        InvocationCache cache = DelegateProxyService.getInvocationCache();
-        return cache.size();
     }
 
     /**
@@ -100,12 +82,6 @@ public final class ApplicationStatus {
         Map<String,Object> section = new LinkedHashMap<>();
         section.put("version", getApplicationVersion());
         status.put("application", section);
-
-        // DMIC
-        section = new LinkedHashMap<>();
-        section.put("size", getDMICSize());
-        section.put("maxSize", getDMICMaxSize());
-        status.put("delegateMethodInvocationCache", section);
 
         // Info cache
         section = new LinkedHashMap<>();